refactor: move chunked Apply methods to Apply.cs
